Fix the line wrapping logic
The logic for when the wrap a line the first time was based on not including
the last name in the line, which doesn't really make sense (I'm not sure why I
did it that way). Now the line is always wrapped if it is longer than the max
line length.